refactor: organize Neuron RAG classes
