+++ src/Domains/Intelligence/Agents/Neuron/BaseKanvasAgent.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Kanvas\Intelligence\Agents\Neuron;
 
-use Kanvas\Intelligence\Agents\Neuron\Concerns\HasKanvasAgentBehavior;
+use Kanvas\Intelligence\Agents\Neuron\Traits\HasKanvasAgentBehavior;
 use Kanvas\Intelligence\Agents\Neuron\Contracts\KanvasNeuronAgent;
 use NeuronAI\Agent\Agent;
