+++ src/Domains/Intelligence/Agents/Neuron/BaseRagAgent.php
@@ -4,13 +4,13 @@ declare(strict_types=1);
 
 namespace Kanvas\Intelligence\Agents\Neuron;
 
-use Kanvas\Intelligence\Agents\Neuron\Concerns\HasKanvasAgentBehavior;
 use Kanvas\Intelligence\Agents\Neuron\Contracts\KanvasNeuronAgent;
+use Kanvas\Intelligence\Agents\Neuron\Traits\HasKanvasAgentBehavior;
 use NeuronAI\RAG\RAG;
 use NeuronAI\Workflow\Node;
 use Override;
 
-class NeuronRagAgent extends RAG implements KanvasNeuronAgent
+class BaseRagAgent extends RAG implements KanvasNeuronAgent
 {
     use HasKanvasAgentBehavior;
