+++ src/Domains/Intelligence/Agents/Neuron/CRM/SalesAgent.php
@@ -8,7 +8,7 @@ use Illuminate\Support\Facades\Blade;
 use Kanvas\Guild\Leads\Models\Lead;
 use Kanvas\Intelligence\Agents\Attributes\AgentTypeDefinition;
 use Kanvas\Intelligence\Agents\Contracts\ConversesWithCustomer;
-use Kanvas\Intelligence\Agents\Neuron\NeuronRagAgent;
+use Kanvas\Intelligence\Agents\Neuron\BaseRagAgent;
 use Kanvas\Intelligence\Agents\Neuron\SalesAssistKanvasMessageHistory;
 use Kanvas\Intelligence\Agents\Neuron\Tools\CRM\ArtifactsTool;
 use Kanvas\Intelligence\Agents\Neuron\Tools\CRM\CalendarEventTool;
@@ -48,7 +48,7 @@ use Override;
     description: 'Conversational sales assistant for inbound prospect chat — qualifies leads, creates them when intent shows, schedules meetings, and recommends inventory using the CRM + inventory tool suite.',
     provider: 'neuron',
 )]
-class SalesAgent extends NeuronRagAgent implements ConversesWithCustomer
+class SalesAgent extends BaseRagAgent implements ConversesWithCustomer
 {
     use HasCustomerPersona;
     use HasLeadRag;
