+++ src/Domains/Intelligence/Agents/Neuron/RAG/Services/RagComponents.php
@@ -2,11 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Kanvas\Intelligence\Agents\Neuron\RAG;
+namespace Kanvas\Intelligence\Agents\Neuron\RAG\Services;
 
 use Baka\Search\SearchEngineResolver;
 use Illuminate\Database\Eloquent\Model;
 use Kanvas\Apps\Models\Apps;
+use Kanvas\Intelligence\Agents\Neuron\RAG\VectorStores\TypesenseVectorStore;
 use Kanvas\Intelligence\Enums\ConfigurationEnum;
 use Kanvas\Intelligence\Knowledge\DataTransferObject\KnowledgeEntity;
 use Kanvas\Intelligence\Knowledge\Enums\LeadRagConfigurationEnum;
